Move get_interface_to_target to network, helping prevent a circular import between pba.py and exploiter modules.
