+++ monkey/infection_monkey/exploit/sshexec.py
@@ -7,10 +7,9 @@ import paramiko
 import infection_monkey.monkeyfs as monkeyfs
 from infection_monkey.exploit import HostExploiter
 from infection_monkey.exploit.tools.helpers import get_target_monkey, get_monkey_depth, build_monkey_commandline
-from infection_monkey.exploit.tools.helpers import get_interface_to_target
 from infection_monkey.model import MONKEY_ARG
+from infection_monkey.network.tools import check_tcp_port, get_interface_to_target
 from infection_monkey.exploit.tools.exceptions import FailedExploitationError
-from infection_monkey.network.tools import check_tcp_port
 from common.utils.exploit_enum import ExploitType
 from common.utils.attack_utils import ScanStatus
 from infection_monkey.telemetry.attack.t1105_telem import T1105Telem
