+++ monkey/infection_monkey/exploit/tools/http_tools.py
@@ -7,7 +7,8 @@ from threading import Lock
 from infection_monkey.network.firewall import app as firewall
 from infection_monkey.network.info import get_free_tcp_port
 from infection_monkey.transport import HTTPServer, LockedHTTPServer
-from infection_monkey.exploit.tools.helpers import try_get_target_monkey, get_interface_to_target
+from infection_monkey.exploit.tools.helpers import try_get_target_monkey
+from infection_monkey.network.tools import get_interface_to_target
 from infection_monkey.model import DOWNLOAD_TIMEOUT
 
 __author__ = 'itamar'
