+++ monkey/infection_monkey/exploit/tools/smb_tools.py
@@ -10,8 +10,9 @@ import infection_monkey.config
 import infection_monkey.monkeyfs as monkeyfs
 from common.utils.attack_utils import ScanStatus
 from infection_monkey.telemetry.attack.t1105_telem import T1105Telem
-from infection_monkey.exploit.tools.helpers import get_interface_to_target
+from infection_monkey.network.tools import get_interface_to_target
 from infection_monkey.config import Configuration
+
 __author__ = 'itamar'
 
 LOG = logging.getLogger(__name__)
